Render two overlapping spheres

+++ include/Sphere.h
@@ -16,7 +16,7 @@ public:
     Sphere() = default;
     Sphere(const std::string& line);
     
-    int mRadius;
+    double mRadius;
     Eigen::Vector3d mPosition;
     Eigen::Vector3d mAmbientReflection;
     Eigen::Vector3d mDiffuseReflection;
